+++ src/Bundle/CustomReport/Controller/Config/UpdateController.php
@@ -17,7 +17,7 @@ use OpenApi\Attributes\JsonContent;
 use OpenApi\Attributes\Put;
 use Pimcore\Bundle\StudioBackendBundle\Bundle\CustomReport\Schema\CustomReportDetails;
 use Pimcore\Bundle\StudioBackendBundle\Bundle\CustomReport\Schema\CustomReportUpdate;
-use Pimcore\Bundle\StudioBackendBundle\Bundle\CustomReport\Service\CustomReportServiceInterface;
+use Pimcore\Bundle\StudioBackendBundle\Bundle\CustomReport\Service\CustomReportConfigServiceInterface;
 use Pimcore\Bundle\StudioBackendBundle\Controller\AbstractApiController;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\NotFoundException;
 use Pimcore\Bundle\StudioBackendBundle\Exception\Api\NotWriteableException;
@@ -43,7 +43,7 @@ final class UpdateController extends AbstractApiController
 
     public function __construct(
         SerializerInterface $serializer,
-        private readonly CustomReportServiceInterface $customReportService
+        private readonly CustomReportConfigServiceInterface $customReportConfigService
     ) {
         parent::__construct($serializer);
     }
@@ -79,6 +79,6 @@ final class UpdateController extends AbstractApiController
         string $name,
         #[MapRequestPayload] CustomReportUpdate $parameters
     ): JsonResponse {
-        return $this->jsonResponse($this->customReportService->updateCustomReport($name, $parameters));
+        return $this->jsonResponse($this->customReportConfigService->updateCustomReport($name, $parameters));
     }
 }
